I changed some small thing. 1. Spacing between notes (to be changed again with measures but I dont like the huge spaces between whole notes) 2. The GUI a bit

+++ src/graphics/MyImages.java
@@ -25,7 +25,7 @@ public class MyImages
 		return img;
 	}
 	
-	public static Image getNoteImage(double beats, boolean flipped)
+	public static Image getNoteImage(int beats, boolean flipped)
 	{
 		//This doesn't correspond to the program:
 		//1 beat = quarter note, 4 = whole note, 1/2 = 1/8th note, usw....
